SR-IOV: improved vsock handshake
Automatically map vsock connections to VF indices by probing the
PF-to-VF interrupt, instead of statically mapping vsock CIDs to VFs.
This allows VFs to be attached to the host system, and also allows
multiple VFs to be attached to a single virtual machine.

Re-enable the SR-IOV tests that were disabled when the vsock transport
was introduced.