cpt: remove double refcount inc
When a communication profile is reused, a refcount is
incremented. However this was done twice, which lead to a leak because
it could never be freed. Eventually nodes that would use that feature
would run out of communication profiles and application couln't start.

Also, move the cxi_tx_profile_dec_refcount() since it applies for
users of the CP not just for the last one. This bug was hidden by the
CP not being freed.

The traces above are using this debug setting:

  echo 8 > /proc/sys/kernel/printk
  echo -n 'file cass_cpt.c +p' >> /proc/dynamic_debug/control

This is test_ucxi_atu using a single CP:

  cxi_ss1 0000:00:08.0: cxi0[] cxi0 cp allocated: tc=BEST_EFFORT tc_type=DEFAULT cp=1 vni_pcp=16 label=3 tc=3 pfq=10 exclusive=0
  cxi_ss1 0000:00:08.0: cxi0[] cxi0 lcid allocated: rgid=1 lcid=0 cp=1
  cxi_ss1 0000:00:08.0: cxi0[] cxi0 lcid freed: rgid=1 lcid=0 cp=1
  cxi_ss1 0000:00:08.0: cxi0[] cxi0 cp freed: cp=1 vni_pcp=16 label=3

After modifying text_ucxi_atu to allocate and free cp2, the CP is not freed, and now leaks:

  cxi_ss1 0000:00:08.0: cxi0[] cxi0 cp allocated: tc=BEST_EFFORT tc_type=DEFAULT cp=1 vni_pcp=10 label=3 tc=3 pfq=10 exclusive=0
  cxi_ss1 0000:00:08.0: cxi0[] cxi0 lcid allocated: rgid=1 lcid=0 cp=1
  cxi_ss1:cxi_cp_alloc:306 Reuse cp:1 rgid:1 lcid:0 refcount:3
  ccess
  cxi_ss1:cxi_cp_free:381 Return cp:1 rgid:1 lcid:0 refcount:2
  cxi_ss1:cxi_cp_free:381 Return cp:1 rgid:1 lcid:0 refcount:1

After removing the extra refcount_inc, the CP is properly freed:

  cxi_ss1 0000:00:08.0: cxi0[] cxi0 cp allocated: tc=BEST_EFFORT tc_type=DEFAULT cp=1 vni_pcp=10 label=3 tc=3 pfq=10 exclusive=0
  cxi_ss1 0000:00:08.0: cxi0[] cxi0 lcid allocated: rgid=1 lcid=0 cp=1
  cxi_ss1:cxi_cp_alloc:305 Reuse cp:1 rgid:1 lcid:0 refcount:2
  ccess
  cxi_ss1:cxi_cp_free:380 Return cp:1 rgid:1 lcid:0 refcount:1
  cxi_ss1 0000:00:08.0: cxi0[] cxi0 lcid freed: rgid=1 lcid=0 cp=1
  cxi_ss1 0000:00:08.0: cxi0[] cxi0 cp freed: cp=1 vni_pcp=10 label=3

NETCASSINI-7674