Prevent enabling SR-IOV in default build
SR-IOV support is a work in progress, and is not intended for end-users
at this point in time. Prevent it from being enabled in the default
build by removing the sriov_configure entrypoint from the pci_driver
struct. The SR-IOV-related sysfs entries will still be populated since
the hardware still adverties SR-IOV support, but the kernel will emit an
appropriate error message if a user attempts to enable SR-IOV.

SR-IOV support code will still build and be present in the driver, but
will not be accessible. To allow SR-IOV to be enabled, invoke the
makefile with ENABLE_SRIOV=1.

Fixes: NETCASSINI-7700

Signed-off-by: Richard Halkyard <richard.halkyard@hpe.com>