Prevent creation of SR-IOV VFs in RPM build
SR-IOV support is a work in progress, and is not intended for end-users
at this point in time. Prevent SR-IOV from being enabled in release
builds by forcing the maximum number of allowed VFs to 0.

SR-IOV support is still enabled by default for development builds, but
RPM builds invoke the makefile with `NO_SRIOV=1`, which will prevent
SR-IOV VF creation in the built driver.

Tests have been modified to skip SR-IOV tests on SR-IOV-disabled builds.

Fixes: NETCASSINI-7700

Signed-off-by: Richard Halkyard <richard.halkyard@hpe.com>